requirements.txt
discord.py>=2.3.2
python-dotenv
motor>=3.3.2
google-generativeai
Pillow
black
nicegui
requests_oauthlib
pyjwt
imageio
